test: 이미지를 삭제하는 공통된 로직을 인터페이스로 분리
