+++ backend/src/test/java/edonymyeon/backend/post/application/PostServiceFindingAllPostsTest.java
@@ -8,19 +8,16 @@ import edonymyeon.backend.image.postimage.repository.PostImageInfoRepository;
 import edonymyeon.backend.member.application.dto.MemberIdDto;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.member.repository.MemberRepository;
+import edonymyeon.backend.post.ImageFileCleaner;
 import edonymyeon.backend.post.application.dto.GeneralFindingCondition;
 import edonymyeon.backend.post.application.dto.GeneralPostInfoResponse;
 import edonymyeon.backend.post.application.dto.PostRequest;
 import edonymyeon.backend.post.application.dto.PostResponse;
 import edonymyeon.backend.post.repository.PostRepository;
-import java.io.File;
-import java.io.FilenameFilter;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
 import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -46,7 +43,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Import(TestConfig.class)
 @TestInstance(Lifecycle.PER_CLASS)
 @DisplayName("게시글 전체 조회 테스트")
-public class PostServiceFindingAllPostsTest {
+public class PostServiceFindingAllPostsTest implements ImageFileCleaner {
 
     public static final String POST_REQUEST1_TITLE = "Lost in Time";
     public static final String POST_REQUEST1_CONTENT = "A young archaeologist discovers a mysterious artifact that transports her back in time, forcing her to navigate ancient civilizations and find a way back home before history unravels.";
@@ -233,19 +230,5 @@ public class PostServiceFindingAllPostsTest {
         );
 
         return postService.createPost(memberId, POST_REQUEST3);
-    }
-
-    @AfterEach
-    public void cleanImageStoreDirectory() {
-        final File targetFolder = new File("src/test/resources/static/img/test_store/");
-        FilenameFilter filter = new FilenameFilter() {
-            @Override
-            public boolean accept(final File dir, final String name) {
-                return !name.equals("test.txt");
-            }
-        };
-        File[] files = targetFolder.listFiles(filter);
-        assert files != null;
-        Arrays.stream(files).forEach(file -> file.delete());
     }
 }
